+++ src/main/kotlin/com/github/com/cakevm/intellij_huff_plugin/ide/inspection/ResolveNameInspection.kt
@@ -1,7 +1,10 @@
 package com.github.com.cakevm.intellij_huff_plugin.ide.inspection
 
 import com.github.com.cakevm.intellij_huff_plugin.language.psi.HuffIncludeDirective
+import com.github.com.cakevm.intellij_huff_plugin.language.psi.HuffMacroConstantReference
+import com.github.com.cakevm.intellij_huff_plugin.language.psi.HuffMacroLabelGoTo
 import com.github.com.cakevm.intellij_huff_plugin.language.psi.HuffVisitor
+import com.github.com.cakevm.intellij_huff_plugin.language.psi.element.base.HuffReferenceElement
 import com.intellij.codeInspection.LocalInspectionTool
 import com.intellij.codeInspection.ProblemHighlightType
 import com.intellij.codeInspection.ProblemsHolder
@@ -19,6 +22,29 @@ class ResolveNameInspection : LocalInspectionTool() {
           }
         }
       }
+
+      override fun visitMacroLabelGoTo(element: HuffMacroLabelGoTo) {
+        checkReference(element) {
+          holder.registerProblem(element, "'${element.identifier.text}' label is undefined", ProblemHighlightType.WARNING)
+        }
+      }
+
+      override fun visitMacroConstantReference(element: HuffMacroConstantReference) {
+        checkReference(element) {
+          holder.registerProblem(element, "'${element.identifier.text}' constant is undefined", ProblemHighlightType.WARNING)
+        }
+      }
+    }
+  }
+}
+
+private fun checkReference(element: HuffReferenceElement, report: () -> Unit) {
+  if (element.reference != null) {
+    // resolve return either 1 reference or null, and because our resolve is not perfect we can return a number
+    // of references, so instead of showing false positives we can use multiResolve
+    val results = element.reference?.multiResolve(false)
+    if (results.isNullOrEmpty()) {
+      report()
     }
   }
 }