+++ src/main/kotlin/com/github/com/cakevm/intellij_huff_plugin/language/psi/element/HuffMacroCallElement.kt
@@ -2,4 +2,4 @@ package com.github.com.cakevm.intellij_huff_plugin.language.psi.element
 
 import com.github.com.cakevm.intellij_huff_plugin.language.psi.element.base.HuffReferenceElement
 
-interface HuffMacroCallElement : HuffReferenceElement {}
+interface HuffMacroCallElement : HuffReferenceElement