+++ src/main/kotlin/com/github/com/cakevm/intellij_huff_plugin/language/psi/element/HuffMacroConstantReferenceElement.kt
@@ -1,9 +1,5 @@
 package com.github.com.cakevm.intellij_huff_plugin.language.psi.element
 
-import com.github.com.cakevm.intellij_huff_plugin.language.psi.element.base.HuffCallable
 import com.github.com.cakevm.intellij_huff_plugin.language.psi.element.base.HuffReferenceElement
 
-interface HuffMacroConstantReferenceElement : HuffReferenceElement {
-
-  fun resolveDefinitions(): List<HuffCallable>?
-}
+interface HuffMacroConstantReferenceElement : HuffReferenceElement