+++ src/main/kotlin/com/github/com/cakevm/intellij_huff_plugin/language/psi/mixin/HuffMacroConstantReferenceMixin.kt
@@ -1,9 +1,7 @@
 package com.github.com.cakevm.intellij_huff_plugin.language.psi.mixin
 
-import com.github.com.cakevm.intellij_huff_plugin.language.psi.HuffIdentifier
 import com.github.com.cakevm.intellij_huff_plugin.language.psi.HuffMacroConstantReference
 import com.github.com.cakevm.intellij_huff_plugin.language.psi.element.HuffMacroLabelGoToElement
-import com.github.com.cakevm.intellij_huff_plugin.language.psi.element.base.HuffCallable
 import com.github.com.cakevm.intellij_huff_plugin.language.psi.impl.HuffNamedElementImpl
 import com.github.com.cakevm.intellij_huff_plugin.language.reference.HuffMacroConstantReferenceReference
 import com.github.com.cakevm.intellij_huff_plugin.language.reference.HuffReference
@@ -13,19 +11,13 @@ import com.intellij.psi.PsiElement
 abstract class HuffMacroConstantReferenceMixin(node: ASTNode) :
   HuffNamedElementImpl(node), HuffMacroLabelGoToElement, HuffMacroConstantReference {
 
-  private fun getReferenceNameElement(expr: HuffIdentifier): PsiElement {
-    return expr
-  }
-
   override val referenceNameElement: PsiElement
-    get() = getReferenceNameElement(this.identifier)
-
-  override fun getReference(): HuffReference = HuffMacroConstantReferenceReference(this as HuffMacroConstantReference)
-
-  override fun resolveDefinitions(): List<HuffCallable>? {
-    return emptyList()
-  }
+    get() = this.identifier
 
   override val referenceName: String
     get() = referenceNameElement.text
+
+  override fun getName(): String? = referenceName
+
+  override fun getReference(): HuffReference = HuffMacroConstantReferenceReference(this as HuffMacroConstantReference)
 }